% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pa_rse.R
\name{create_pa_rse}
\alias{create_pa_rse}
\title{Create a RangedSummarizedExperiment from recount3}
\usage{
create_pa_rse(
  organism = c("human", "mouse"),
  project = NULL,
  annotation = NULL,
  sample_id = NULL,
  bed_cols = c("chr", "start", "end", "ID", "score", "strand")
)
}
\arguments{
\item{organism}{A character(1) specifying which organism you want to download
data from. Supported options are "human" or "mouse".}

\item{project}{A character(1) with the ID for a given study. See
\link[recount3:available_projects]{available_projects} from the recount3 package.}

\item{annotation}{A character(1) with the path to a BED file containing
the coordinates of annotated PA-sites.}

\item{sample_id}{A character vector with the ID(s) for a given sample inside
the study. See
\link[recount3:available_samples]{available_samples} from the recount3 package.}

\item{bed_cols}{A character vector with the column names for the BED file.
By default it uses the standard fields for a BED file ("chr", "start",
"end", "ID", "score", "strand").}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
`create_pa_rse creates a RangedSummarizedExperiment object by extracting
expression information from recount3 bigWig files from annotated polyA
sites provided through a BED file. It will also pull relevant metadata
for the project through recount3.
}
\examples{
## Explore available projects in recount3
human_projects <- recount3::available_projects()

## Find the project you are interested in and obtain PA expression
create_pa_rse(organism = "mouse", project="SRP048707", annotation="mysites.bed")

}
\author{
Eddie Imada
}
