% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_counts.R
\name{ilr_counts}
\alias{ilr_counts}
\title{Obtain an isometric log ratio transformed matrix from REPAC results}
\usage{
ilr_counts(se, results)
}
\arguments{
\item{se}{The (Ranged)SummarizedExperiment object used on the analysis.}

\item{results}{A data.frame or tibble with results from REPAC}
}
\value{
An isometric log ratio transformed matrix.
}
\description{
\code{ilr_counts} generates an sometric log ratio transformed matrix based on the
References and Sites from a REPAC result.
}
\examples{
## fit model and test for DPU
groups <- rep(c("A", "B"), each=5)
dMat <- model.matrix(~ 0+groups)
colnames(dMat) <-  gsub("groups", "",colnames(dMat))
cMat <- makeContrasts(
    levels=colnames(dMat),
    BvsA= B - A
)
results <- fit_repac(se, dMat, cMat)
results <- resolve_comparisons(results)
mat <- ilr_counts(se, results$BvsA)
}
\author{
Eddie Imada
}
