% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse_from_bam.R
\name{rse_from_bam}
\alias{rse_from_bam}
\title{Create a RangedSummarizedExperiment from bam files}
\usage{
rse_from_bam(bams, annotation, colData, ...)
}
\arguments{
\item{bams}{A character list with the path for the bam files.}

\item{annotation}{A GenomicRanges object with the coordinates of annotated
PA-sites. This object must have an elementMetadata column named "gene_name" with
gene-level IDs.}

\item{colData}{A data frame containing sample metadaata}

\item{...}{Additional parameters passed to \code{Rsubread::featureCounts}}
}
\value{
A
\link[SummarizedExperiment:RangedSummarizedExperiment-class]{RangedSummarizedExperiment-class}
object.
}
\description{
`rse_from_bam creates a RangedSummarizedExperiment object by extracting
counts from bam files of annotated polyA sites provided through an annotation.
}
\examples{
## Get path to bam files
bams <- list.file("~", full.names=T, pattern = "*.bam$")

## Load and quantify human hg38 PAS sites
data(hg38_pa)
pheno <- data.frame(groups=rep(c("A", "B"), each=5))
rse_from_bam(bams, hg38_pa, pheno, strandSpecific = 0,  allowMultiOverlap = F,
fraction = F, countMultiMappingReads=F, isPairedEnd = T, nthreads = 8,
countReadPairs = T,  requireBothEndsMapped = F)

}
\author{
Eddie Imada
}
