% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_repac.R
\name{fit_repac}
\alias{fit_repac}
\title{Fit model and test for differential polyA usage (DPU)}
\usage{
fit_repac(se, design, contrasts = NULL)
}
\arguments{
\item{se}{A (Ranged)SummarizedExperiment object.}

\item{design}{A design matrix indicating the groups
to be compared. See \link[stats:model.matrix]{model.matrix}.}

\item{contrasts}{A numeric matrix with rows corresponding to coefficients in the
design matrix and columns containing contrasts. See \link[limma:makeConstrasts]{makeContrasts}.}
}
\value{
A list of tibbles with length equal the number of contrasts.
Each tibble contains the following columns:
gene_name - Name of the gene associated with the PAS (e.g., SYMBOL)
Ref - ID of the reference PAS
ID - ID of the PAS tested in comparison to \code{Ref}
cFC - Compositional fold-change in the ilr-space
mcDiff - Mean difference between the coverages in percentage across groups
t - T-statistic value
p.val - p-value
}
\description{
\code{fit_repac} transforms the counts obtained from \code{create_pa_rse()} into compositions
and performs an isometric log ratio transformation prior to fitting the model
}
\examples{
## fit model and test for DPU
groups <- rep(c("A", "B"), each=5)
dMat <- model.matrix(~ 0+groups)
colnames(dMat) <-  gsub("groups", "",colnames(dMat))
cMat <- makeContrasts(
    levels=colnames(dMat),
    BvsS= B - A
)
results <- fit_repac(se, dMat, cMat)

}
\author{
Eddie Imada
}
